split menu tests
